Add DistSQLBackendHandler.
