+++ proxy/backend/src/main/resources/META-INF/services/org.apache.shardingsphere.distsql.handler.resultset.DistSQLResultSet
@@ -15,7 +15,7 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.proxy.backend.handler.distsql.rql.resource.StorageUnitResultSet
+org.apache.shardingsphere.proxy.backend.handler.distsql.rql.storage.unit.StorageUnitResultSet
 org.apache.shardingsphere.proxy.backend.handler.distsql.rql.rule.LogicalTableResultSet
 org.apache.shardingsphere.proxy.backend.handler.distsql.rql.rule.SingleTableResultSet
 org.apache.shardingsphere.proxy.backend.handler.distsql.rql.rule.SingleTableRuleResultSet
